backend/requirements.txt
fastapi
uvicorn
tortoise-orm
pykakasi
asyncpg